Name Suggestion
The Peach Consortium
^
This is a name suggestion. I like the sound of it, it's official but cute. I like that it can be written like this The :peach: Consortium and I like that it also enables us to inofficially still be called buttsortium. I like the name buttsortium because it's fun and makes things a bit lighter. I don't want this space to only feel like work (I like playfulness!) while it would simultaniously be great if the name can be a good official face outwards.

I still think we can use the domain name André registered for this year and then switch once the time comes, or set up dual link to a site with two domains.